frontend/requirements.txt
Pillow
streamlit==1.2.0
jina[client]==3.0.0
